% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_thresholds}
\alias{df_thresholds}
\title{Displacement thresholds data frame}
\format{
A data frame with 3 rows and 4 columns:
\describe{
\item{weekly}{Weekly thresholds.}
\item{monthly}{Monthly thresholds.}
\item{quarterly}{Quarterlythresholds.}
\item{yearly}{Yearly thresholds.}
}
}
\usage{
df_thresholds
}
\description{
A data frame of global thresholds for flagging different types of
displacement (Conflict, Disaster, and Other) across different periods of
time (weekly, monthly, quarterly, and yearly). Thresholds were set to
be a clear whole number near to but below the 95th percentile of data
calculated for all data uploaded in the database at the end of 2022.
}
\details{
Used with \code{\link[=idmc_flagging]{idmc_flagging()}} to generate the global flags that do not rely on
country-level anomalies.
}
\keyword{datasets}
