% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_thresholds_min}
\alias{df_thresholds_min}
\title{Displacement threshold minimum data frame}
\format{
A data frame with 1 rows and 4 columns:
\describe{
\item{weekly}{Weekly thresholds.}
\item{monthly}{Monthly thresholds.}
\item{quarterly}{Quarterlythresholds.}
\item{yearly}{Yearly thresholds.}
}
}
\usage{
df_thresholds_min
}
\description{
A data frame of global minimums for flagging across different periods of
time (weekly, monthly, quarterly, and yearly). Thresholds were set based on
agreement with CERF to reduce the total number of flags.
}
\details{
Used with \code{\link[=idmc_flagging]{idmc_flagging()}} to generate the global flags that do not rely on
country-level anomalies. Displacement for that time frame below the threshold
minimum will not flag, even if it is greater than the country anomaly
threshold.
}
\keyword{datasets}
