% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_get_data.R
\name{idmc_get_data}
\alias{idmc_get_data}
\title{Get data from the IDMC API}
\usage{
idmc_get_data()
}
\value{
Tibble of displacement data
}
\description{
\code{idmc_get_data()} calls the IDMC API to retrieve displacement data. The data
is converted from JSON to a data frame, date columns converted to \code{Date}
types, and returned as a \link[tibble:tibble]{tibble::tibble}.
}
