% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_rolling_sum.R
\name{idmc_rolling_sum}
\alias{idmc_rolling_sum}
\title{Create rolling sums of displacement data}
\usage{
idmc_rolling_sum(df, infill = TRUE)
}
\arguments{
\item{df}{Event displacement data frame, generated from \code{\link[=idmc_get_data]{idmc_get_data()}}.}

\item{infill}{If \code{TRUE}, the default, days with no displacement are infilled
with \code{0}. If \code{FALSE}, no infilling is done automaticaly. However, without
infilling, rolling sums will not be accurate, so be sure to infill on
on your own.}
}
\description{
Create rolling sums of displacement data
}
