% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_transform_daily.R
\name{idmc_transform_daily}
\alias{idmc_transform_daily}
\title{Transform displacement event data to daily data}
\usage{
idmc_transform_daily(df)
}
\arguments{
\item{df}{Event displacement data frame, generated from \code{\link[=idmc_get_data]{idmc_get_data()}}.}
}
\value{
Data frame of daily displacement by country and displacement type.
}
\description{
\code{idmc_transform_daily()} transforms event data from the IDMC API (accessed
through \code{\link[=idmc_get_data]{idmc_get_data()}}). The data for each event is spread out between
the start and end date, with the total displacement uniformly distributed
across all days. For each country and displacement type (conflict, disaster,
or other), all displacement on a day is summed up to create a total
daily displacement figure.
}
