% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc-package.R
\docType{package}
\name{idmc-package}
\alias{idmc}
\alias{idmc-package}
\title{idmc: Load and Wrangle IDMC Displacement Data}
\description{
idmc provides easy access to event displacement data available from the Internal Displacement Monitoring Centre. Convenient functions are available to load and wrangle the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocha-dap/idmc}
  \item Report bugs at \url{https://github.com/ocha-dap/idmc/issues}
}

}
\author{
\strong{Maintainer}: Seth Caldwell \email{caldwellst@gmail.com} [complainant-appellant]

}
\keyword{internal}
