% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_flagging.R
\name{idmc_flagging}
\alias{idmc_flagging}
\title{Flag IDMC displacement data}
\usage{
idmc_flagging(df)
}
\arguments{
\item{df}{Event displacement data frame, generated from \code{\link[=idmc_rolling_sum]{idmc_rolling_sum()}}.}
}
\description{
\code{idmc_flagging()} generates flags on IDMC displacement data. Flags can be
generated based country-level anomalies, global thresholds, and if the
displacement is the first reported displacement in a period of time. The
input dataset should be generated from \code{\link[=idmc_rolling_sum]{idmc_rolling_sum()}}. All flags are
calculated separately on displacement type in the IDMC database: conflict,
displacement, and other.
}
\details{
Flags are generated as:
\itemize{
\item Country-level flags. Anomalies are generated based
on the values being in the top \%5 of observed displacement values for that
country (excluding 0, days without displacement). These flags are generated
across time spans from daily to yearly to detect anomalies at different
levels. Anomalies are calculated separately for each type of displacement
(conflict, disaster, other) to ensure that new shifts in conflict
displacement dynamics in a country like Mozambique are not missed due to
previous large displacements due to natural disasters.
\item Global flags based on static thresholds. These are calculated for total
daily displacement in a country from daily to yearly to find overall
anomalies. The thresholds are set based on the quantiles for displacement
data in the database by the end of 2022. Details in
\link{df_thresholds_lim}.
\item Flags if there is displacement for the first time in 3 months, 6 months,
or one year.
}

For country-level and global flags, a minimum number of individuals are
needed across all timepoints. This is defined in \link{df_thresholds_min}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
idmc_get_data() \%>\%
  idmc_transform_daily() \%>\%
  idmc_rolling_sum() \%>\%
  idmc_flagging()
\dontshow{\}) # examplesIf}
}
