% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_get_data.R
\name{idmc_get_data}
\alias{idmc_get_data}
\title{Get data from the IDMC API}
\usage{
idmc_get_data(api_url = NULL)
}
\arguments{
\item{api_url}{URL to the IDMC API. If \code{NULL}, the default, searches for
the \code{IDMC_API} environment variable.}
}
\value{
Tibble of displacement data. Description of the data frame variables
are included in the documentation for the
\href{https://www.internal-displacement.org/sites/default/files/IDMC_IDU_API_Codebook_14102020.pdf}{IDMC IDU API}. # nolint
}
\description{
\code{idmc_get_data()} calls the IDMC API to retrieve displacement data. The data
is converted from JSON to a data frame, date columns converted to \code{Date}
types, and returned as a \link[tibble:tibble]{tibble::tibble}. A valid URL endpoint must be
saved in your environment as \code{IDMC_API}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
idmc_get_data()
\dontshow{\}) # examplesIf}
}
